Update README.md
I don't know if this is accurate, but I was not otherwise clear on whether the queries were relative to root or whether they could be found anywhere. I assume `recurse` allows them to be found anywhere, whereas otherwise paths must be absolute. I am also not clear on the definition of `recurse`: "looking in children of the object as objects in arrays in the given property value". Is there a means to distinguish absolute (or relative) paths by looking, e.g., for the third array item in an array which is the value of an object property "myKey" anywhere in the document, or the object which is the value of the "key2" property which is on an object with the "key1" property relative to root. It would really help if you could make more clear what is going on with hierarchical paths. Thanks!